priceとoldカラムを修正

+++ db/migrate/20191224064554_create_houses.rb
@@ -2,9 +2,9 @@ class CreateHouses < ActiveRecord::Migration[5.2]
   def change
     create_table :houses do |t|
       t.string :name
-      t.integer :price
+      t.float :price
       t.string :address
-      t.integer :old
+      t.float :old
       t.text :remarks
 
       t.timestamps
